requirements.txt
playwright==1.40.0
python-dotenv==1.0.0
schedule==1.2.0
openpyxl==3.1.2
